added routing, fixed standard, refactoring code, is LoggedIn logic implemented
